requirements.txt
pymongo==4.8.0
minio==7.2.7
requests==2.32.3
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
python-dateutil==2.9.0.post0